+++ src/main/java/org/example/employee_performance_review_api/domain/feature/CalculateEmployeeBonus.java
@@ -2,8 +2,9 @@ package org.example.employee_performance_review_api.domain.feature;
 
 import org.example.employee_performance_review_api.domain.model.employee.Employee;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface CalculateEmployeeBonus {
-    double handle(Employee employee);
+    BigDecimal handle(Employee employee);
 }
