+++ src/main/java/org/example/employee_performance_review_api/domain/model/employee/EmployeeModelBuilder.java
@@ -7,6 +7,7 @@ import org.example.employee_performance_review_api.domain.model.employee.enums.T
 import org.example.employee_performance_review_api.domain.validator.ModelValidator;
 
 import javax.inject.Named;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -15,14 +16,14 @@ import java.util.UUID;
 public class EmployeeModelBuilder {
     private final ModelValidator modelValidator;
 
-    public Employee build(String firstName, String lastName, Team team, PerformanceRating rating, EmployeeType employeeType, double yearlySalary, int employmentTime, List<Double> yearlyBonuses) {
+    public Employee build(String firstName, String lastName, Team team, PerformanceRating rating, EmployeeType employeeType, BigDecimal yearlySalary, int employmentTime, List<BigDecimal> yearlyBonuses) {
 
         return modelValidator.validate(new Employee(UUID.randomUUID(), firstName, lastName, team, rating, employeeType, yearlySalary, employmentTime, yearlyBonuses));
 
         // TODO: Fix this place to validate the user input.
     }
 
-    public Employee build(UUID id, String firstName, String lastName, Team team, PerformanceRating rating, EmployeeType employeeType, double yearlySalary, int employmentTime, List<Double> yearlyBonuses) {
+    public Employee build(UUID id, String firstName, String lastName, Team team, PerformanceRating rating, EmployeeType employeeType, BigDecimal yearlySalary, int employmentTime, List<BigDecimal> yearlyBonuses) {
 
         return modelValidator.validate(new Employee(id, firstName, lastName, team, rating, employeeType, yearlySalary, employmentTime, yearlyBonuses));
